requirements.txt
sympy== 1.10 
seaborn==0.11.2
requests==2.22.0  
matplotlib==3.5.1 
networkx==2.7.1  
pyTelegramBotAPI==4.4.0
Pillow==7.0.0
